+++ Sources/fonts.swift
@@ -11,12 +11,10 @@ import StencilSwiftKit
 import SwiftGenKit
 
 let fontsCommand = command(
-  outputOption,
-  templateOption(prefix: "fonts"), templatePathOption,
+  outputOption, templateNameOption, templatePathOption, paramsOption,
   Option<String>("enumName", "", flag: "e", description: "The name of the enum to generate (DEPRECATED)"),
-  VariadicOption<String>("param", [], description: "List of template parameters"),
   VariadicArgument<Path>("DIR", description: "Directory to parse.", validator: dirsExist)
-  ) { output, templateName, templatePath, enumName, parameters, paths in
+  ) { output, templateName, templatePath, parameters, enumName, paths in
     // show error for old deprecated option
     guard enumName.isEmpty else {
       throw TemplateError.deprecated(option: "enumName", replacement: "Please use '--param enumName=...' instead.")
@@ -28,7 +26,7 @@ let fontsCommand = command(
       }
 
       let templateRealPath = try findTemplate(
-        prefix: "fonts", templateShortName: templateName, templateFullPath: templatePath
+        subcommand: "fonts", templateShortName: templateName, templateFullPath: templatePath
       )
 
       let template = try StencilSwiftTemplate(templateString: templateRealPath.read(),