+++ Sources/strings.swift
@@ -10,12 +10,10 @@ import StencilSwiftKit
 import SwiftGenKit
 
 let stringsCommand = command(
-  outputOption,
-  templateOption(prefix: "strings"), templatePathOption,
+  outputOption, templateNameOption, templatePathOption, paramsOption,
   Option<String>("enumName", "", flag: "e", description: "The name of the enum to generate (DEPRECATED)"),
-  VariadicOption<String>("param", [], description: "List of template parameters"),
   Argument<Path>("FILE", description: "Localizable.strings file to parse.", validator: fileExists)
-) { output, templateName, templatePath, enumName, parameters, path in
+) { output, templateName, templatePath, parameters, enumName, path in
   // show error for old deprecated option
   guard enumName.isEmpty else {
     throw TemplateError.deprecated(option: "enumName", replacement: "Please use '--param enumName=...' instead.")
@@ -26,7 +24,7 @@ let stringsCommand = command(
     try parser.parseFile(at: path)
 
     let templateRealPath = try findTemplate(
-      prefix: "strings", templateShortName: templateName, templateFullPath: templatePath
+      subcommand: "strings", templateShortName: templateName, templateFullPath: templatePath
     )
     let template = try StencilSwiftTemplate(templateString: templateRealPath.read(),
                                             environment: stencilSwiftEnvironment())