+++ Sources/templates.swift
@@ -19,54 +19,54 @@ let templatesListCommand = command(
   },
   outputOption
 ) { onlySubcommand, output in
-  let customTemplates = (try? appSupportTemplatesPath.children()) ?? []
-  let bundledTemplates = (try? bundledTemplatesPath.children()) ?? []
   var outputLines = [String]()
 
-  let printTemplates = { (prefix: String, list: [Path]) in
-    for file in list where file.lastComponent.hasPrefix("\(prefix)-") && file.extension == "stencil" {
+  let printTemplates = { (subcommand: String, path: Path) in
+    guard let files = try? (path + subcommand).children() else { return }
+
+    for file in files where file.extension == "stencil" {
       let basename = file.lastComponentWithoutExtension
-      let idx = basename.index(basename.startIndex, offsetBy: prefix.characters.count+1)
-      let name = basename[idx..<basename.endIndex]
-      outputLines.append("   - \(name)")
+      outputLines.append("   - \(basename)")
     }
   }
 
   let subcommandsToList = onlySubcommand.isEmpty ? allSubcommands : [onlySubcommand]
-  for prefix in subcommandsToList {
-    outputLines.append("\(prefix):")
+  for subcommand in subcommandsToList {
+    outputLines.append("\(subcommand):")
     outputLines.append("  custom:")
-    printTemplates(prefix, customTemplates)
+    printTemplates(subcommand, appSupportTemplatesPath)
     outputLines.append("  bundled:")
-    printTemplates(prefix, bundledTemplates)
+    printTemplates(subcommand, bundledTemplatesPath)
   }
 
   outputLines.append("---")
   outputLines.append("You can add custom templates in \(appSupportTemplatesPath).")
-  outputLines.append("Simply name them 'subcmd-customname.stencil' where subcmd is one of the swiftgen subcommand,")
-  outputLines.append("namely " + allSubcommands.map({"\($0)-xxx.stencil"}).joined(separator: ", ") + ".")
+  outputLines.append("Simply place them in a subfolder of this path named after the corresponding swiftgen")
+  outputLines.append("subcommand, namely " + allSubcommands.map({"'\($0)'"}).joined(separator: ", ") + ". Your")
+  outputLines.append("template must be in the right subfolder, and must have the extension `.stencil`.")
+  outputLines.append("")
+  outputLines.append("For example, if you want to create a custom template named \"customname\" for the strings")
+  outputLines.append("command, place it in \"\(appSupportTemplatesPath)/strings/customname.stencil\".")
   outputLines.append("")
 
   output.write(content: outputLines.joined(separator: "\n"))
 }
 
+// Defines a 'generic' command for doing an operation on a named template. It'll receive the following
+// arguments from the user:
+// - 'subcommand'
+// - 'template'
+// These will then be converted into an actual template path, and passed to the result closure.
 private func templatePathCommandGenerator(execute: @escaping (Path, OutputDestination) throws -> Void) -> CommandType {
   return command(
-    Argument<String>("name",
-      description: "the name of the template to find, like `colors` for the default one" +
-      " or `colors-rawValue' for a specific one"),
+    Argument<String>("subcommand",
+      description: "the name of the subcommand for the template, like `colors`"),
+    Argument<String>("template",
+      description: "the name of the template to find, like `swift3` or `dot-syntax`"),
     outputOption
-  ) { name, output in
+  ) { subcommand, name, output in
     do {
-      let (prefix, shortName): (String, String)
-      if let hyphenPos = name.characters.index(of: "-") {
-        prefix = String(name.characters[name.characters.startIndex..<hyphenPos])
-        shortName = String(name.characters[name.characters.index(after: hyphenPos)..<name.characters.endIndex])
-      } else {
-        prefix = name
-        shortName = "default"
-      }
-      let path = try findTemplate(prefix: prefix, templateShortName: shortName, templateFullPath: "")
+      let path = try findTemplate(subcommand: subcommand, templateShortName: name, templateFullPath: "")
       try execute(path, output)
     } catch {
       printError(string: "error: failed to read template: \(error)")